Refactor rpc_doclayout to use Strategy Pattern
- Unified `rpc_doclayout1.py` through `rpc_doclayout7.py` into a single `rpc_doclayout.py` using a Strategy Pattern.
- Implemented `LayoutStrategy` abstract base class and concrete strategies for V1-V7.
- Updated `babeldoc/main.py` to use `--layout-version` argument instead of multiple `--rpc-doclayoutX` arguments.
- Deleted redundant `rpc_doclayout{2..7}.py` files.
- Ensures maintainability and cleaner CLI interface.
